+++ editor/assets/shaders/mouse-picking-selector.frag
@@ -1,7 +1,8 @@
 #version 460
+#extension GL_ARB_gpu_shader_int64 : require
 #extension GL_EXT_scalar_block_layout : require
 
-layout(location = 0) in flat uint outEntity;
+layout(location = 0) in flat uint64_t outEntity;
 
 #include "mouse-picking-base.glsl"
